feat: pool swaps
