--- src/components/contextual/pages/pool/PoolActivitiesCard/types.ts
@@ -1,4 +1,0 @@
-export enum PoolActivityTab {
-  ALL_ACTIVITY = 'allActivity',
-  USER_ACTIVITY = 'userActivity'
-}
