+++ src/components/contextual/pages/pool/PoolTransactionsCard/types.ts
@@ -1,0 +1,5 @@
+export enum PoolTransactionsTab {
+  ALL_ACTIVITY = 'allActivity',
+  USER_ACTIVITY = 'userActivity',
+  SWAPS = 'swaps'
+}
