+++ src/components/contextual/pages/pool/index.ts
@@ -2,5 +2,5 @@ export { default as PoolChart } from './PoolChart.vue';
 export { default as PoolStatCards } from './PoolStatCards.vue';
 export { default as PoolBalancesCard } from './PoolBalancesCard.vue';
 export { default as PoolActionsCard } from './PoolActionsCard.vue';
-export { default as PoolActivitiesCard } from './PoolActivitiesCard/PoolActivitiesCard.vue';
+export { default as PoolTransactionsCard } from './PoolTransactionsCard/PoolTransactionsCard.vue';
 export { default as MyPoolBalancesCard } from './MyPoolBalancesCard.vue';
