+++ src/composables/queries/usePoolUserActivitiesQuery.ts
@@ -5,7 +5,7 @@ import { UseInfiniteQueryOptions } from 'react-query/types';
 import QUERY_KEYS from '@/constants/queryKeys';
 import { POOLS } from '@/constants/pools';
 
-import BalancerSubgraph from '@/services/balancer/subgraph/balancer-subgraph.service';
+import { balancerSubgraphService } from '@/services/balancer/subgraph/balancer-subgraph.service';
 import { PoolActivity } from '@/services/balancer/subgraph/types';
 import useWeb3 from '@/services/web3/useWeb3';
 import useNetwork from '../useNetwork';
@@ -19,9 +19,6 @@ export default function usePoolUserActivitiesQuery(
   id: string,
   options: UseInfiniteQueryOptions<UserPoolActivitiesQueryResponse> = {}
 ) {
-  // SERVICES
-  const balancerSubgraph = new BalancerSubgraph();
-
   // COMPOSABLES
   const { account, isWalletReady } = useWeb3();
   const { networkId } = useNetwork();
@@ -38,7 +35,7 @@ export default function usePoolUserActivitiesQuery(
 
   // METHODS
   const queryFn = async ({ pageParam = 0 }) => {
-    const poolActivities = await balancerSubgraph.poolActivities.get({
+    const poolActivities = await balancerSubgraphService.poolActivities.get({
       first: POOLS.Pagination.PerPage,
       skip: pageParam,
       where: {
